+++ cfm/cfm-code/EmNNModel.h
@@ -49,7 +49,7 @@ public:
 
   double getRegularizationTerm(unsigned int energy) override;
 
-  void updateGradientForRegularizationTerm(double *grads,
+  void updateGradientForRegularizationTerm(float *grads,
                                            unsigned int energy) override;
 
   void writeParamsToFile(std::string &filename) override;
